Added Picture-in-Picture support to the video example fragment
Added pip action to the video player in Video Consumption Examples
fragment. Upon clicking on this action, the video player opens in pip
mode where the user can navigate other activities, while the video plays
in a corner of the screen.
Also, made the video visible in the background when home screen button
is pressed on a currently playing video.