Added video grid cards example with PIP-enabled functionalty
Added a new card to the home screen that displays videos in a vertical
grid. The videos are fetched from a json URL and rendered using
ImageCardView presenter defined for a new VideoCard data model. Clicking
on each video card will launch a fresh instance of the playback activity.
Videos can be picture-in-pictured (PIP). For example, we can open one
video in PIP mode, while playing another one in full-screen.