Fixed pip media control button issues, made pip singleTask, added NowPlayingCard for videos and added MediaSession activity for Music and Videos
Pip play/pause control buttons weren't displayed which is now fixed. The
media players for both the videos and music service are now resposive to
media control buttons on the remote from anywhere outside the app. Added
NowPlayingCard for the currently active video item to the launcher.
Added MediaSession intent to be launched when clicking on
NowPlayingCard. Made PiP a singleTask (for now) so that no two videos
play at the same time.