Used support fragments, activities, and MediaSessionCompat to make the app compatible for API 17 and above. Added animation for MediaItemList. Added support for fast-forward and rewind actions for playback.
All the playback fragments are now using support library v4 fragments and
FragmentActivity. This makes it possible for the app to run on
pre-lollipop platforms. Provided support for fast-forward and rewind actions
in MediaPlayer for API 17 and above. Added animated peak bars for
currently playing media items that were introduced in support library
24. Did a number of refactoring to MediaPlayerGlue.